requirements.txt
django>=5.0
djangorestframework>=3.14
celery>=5.3
redis>=5.0
pypdf2>=3.0
chromadb>=0.4
langchain>=0.1
langchain-google-genai>=0.0.1
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
psycopg2-binary>=2.9
langchain-community>=0.0.1
django-storages>=1.14
boto3>=1.28
chromadb-client>=0.4
django-redis>=5.4
drf-extensions>=0.7
requests>=2.28
